Generate more coherent text.
Randomly choosing a starting n-gram unconditionally most often generates
output text that is not a sensible start of a sentence. Therefore, this
change randomly chooses a starting n-gram until the chosen n-gram
represents the end of a sentence, thus meaning that the first word
chosen is a word that follows then end of a sentence, and therefore must
be a word that starts a sentence.

Further, to end the generated text at the end of a sentence, this change
also continues to choose words until the currently constructed n-gram
represents the end of a sentence, which means that the most recently
generated word ends a sentence.